The connexion of a user is now enabled. The LoginAuthenticator must authenticate a user with his login and password. The JWTRefreshAuthenticator must check if the refresh token was sent by the user and if the access token is expired. If he's expired, the authenticator create another pair of tokens. The JWTAuthenticator must check if the credentials present in the access token are valid. If it's the case the operation ordered by the user will be done. If it's not the case, the operation is refused.
For the authentication, there are two different entry point: the first is the LoginAuthenticator and the second is the JWTRefreshAuthenticator. If the refresh token is expired, impossible for the user to continue his operation.

To finish the logout method in the SecurityController has been modified, now the cookies are erased of the browser of the client and the refresh token stored in the database is erase.